Add comments for resolve issue when rerendiring Provider
On tests we can't just rerender Get component but rerender Provider and
Get all together. This makes Provider to create a new defaultResolve
function on every rerender.

To mock a real scenario, resolve function is created in tests and passed
as property for Provider so we can keep same the instance on every
rerender, which is what it would happen on a real app.