Fix build (babel helper-compilation-targets workaround)
This is a bug in babel: https://github.com/babel/babel/issues/11216
It is closed but points to an issue in node: https://github.com/nodejs/node/issues/32107
The issue in node is closed as not an issue, so it seems as though babel needs to fix.

Also, check that build works in PRs.